Refactor: extract policy and consumer packages; move jsonschema to internal/ and update imports
