+++ internal/jsonschema/errors.go
