+++ router_test.go
@@ -1,14 +1,15 @@
 package sqsrouter
 
 import (
-	"context"
-	"errors"
-	"fmt"
-	"sync"
-	"testing"
+    "context"
+    "errors"
+    "fmt"
+    "sync"
+    "testing"
 
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
+    "github.com/stretchr/testify/assert"
+    "github.com/stretchr/testify/require"
+    "github.com/hatsunemiku3939/sqsrouter/policy"
 )
 
 const (
@@ -149,26 +150,25 @@ func TestRouter_Route(t *testing.T) {
 		assert.True(t, result.HandlerResult.ShouldDelete)
 	})
 
-	t.Run("policy can override handler error decision", func(t *testing.T) {
-		// Custom policy that forces retry on handler errors regardless of handler's ShouldDelete
-		tp := &testPolicy{
-			decideFunc: func(ctx context.Context, st *RouteState, kind FailureKind, inner error, current RoutedResult) RoutedResult {
-				if kind == FailHandlerError {
-					current.HandlerResult.ShouldDelete = false
-					if inner != nil && current.HandlerResult.Error == nil {
-						current.HandlerResult.Error = inner
-					}
-					return current
-				}
-				return current
-			},
-		}
-		r, err := NewRouter(testEnvelopeSchema, WithPolicy(tp))
-		require.NoError(t, err)
-		// Handler asks to delete even on error
-		r.Register(testMessageType, testMessageVersion, func(_ context.Context, _, _ []byte) HandlerResult {
-			return HandlerResult{ShouldDelete: true, Error: errors.New("boom")}
-		})
+    t.Run("policy can override handler error decision", func(t *testing.T) {
+        // Custom policy that forces retry on handler errors regardless of handler's ShouldDelete
+        tp := policy.Policy(policy.ImmediateDeletePolicy{})
+        // Wrap ImmediateDeletePolicy with a decorator behavior for this test
+        tp = policy.Policy(policyFunc(func(ctx context.Context, kind policy.FailureKind, inner error, current policy.Result) policy.Result {
+            if kind == policy.FailHandlerError {
+                current.ShouldDelete = false
+                if inner != nil && current.Error == nil {
+                    current.Error = inner
+                }
+            }
+            return current
+        }))
+        r, err := NewRouter(testEnvelopeSchema, WithPolicy(tp))
+        require.NoError(t, err)
+        // Handler asks to delete even on error
+        r.Register(testMessageType, testMessageVersion, func(_ context.Context, _, _ []byte) HandlerResult {
+            return HandlerResult{ShouldDelete: true, Error: errors.New("boom")}
+        })
 
 		payload := `{"userId": "123", "username": "test"}`
 		msg := createTestMessage(t, testMessageType, testMessageVersion, payload)
@@ -270,6 +270,13 @@ func TestRouter_Route(t *testing.T) {
 		assert.False(t, called, "handler must not be invoked when metadata unmarshal fails")
 	})
 
+}
+
+// policyFunc allows using a function as a policy.Policy for tests.
+type policyFunc func(ctx context.Context, kind policy.FailureKind, inner error, current policy.Result) policy.Result
+
+func (f policyFunc) Decide(ctx context.Context, kind policy.FailureKind, inner error, current policy.Result) policy.Result {
+    return f(ctx, kind, inner, current)
 }
 
 func TestRouter_Concurrency(t *testing.T) {
