+++ test/e2e/main.go
@@ -1,18 +1,20 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"log"
-	"os"
+    "context"
+    "encoding/json"
+    "fmt"
+    "log"
+    "os"
 	"os/signal"
 	"syscall"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/service/sqs"
-	"github.com/hatsunemiku3939/sqsrouter"
+    "github.com/aws/aws-sdk-go-v2/aws"
+    "github.com/aws/aws-sdk-go-v2/config"
+    "github.com/aws/aws-sdk-go-v2/service/sqs"
+    "github.com/hatsunemiku3939/sqsrouter"
+    "github.com/hatsunemiku3939/sqsrouter/consumer"
+    "github.com/hatsunemiku3939/sqsrouter/policy"
 )
 
 const (
@@ -104,15 +106,14 @@ func E2EMiddleware() sqsrouter.Middleware {
 type forceRetryOnHandlerErr struct{}
 
 // Decide implements the Policy interface for the custom behavior.
-func (forceRetryOnHandlerErr) Decide(_ context.Context, _ *sqsrouter.RouteState, kind sqsrouter.FailureKind, inner error, rr sqsrouter.RoutedResult) sqsrouter.RoutedResult {
-	if kind == sqsrouter.FailHandlerError {
-		rr.HandlerResult.ShouldDelete = false
-		if inner != nil && rr.HandlerResult.Error == nil {
-			rr.HandlerResult.Error = inner
-		}
-		return rr
-	}
-	return rr
+func (forceRetryOnHandlerErr) Decide(_ context.Context, kind policy.FailureKind, inner error, current policy.Result) policy.Result {
+    if kind == policy.FailHandlerError {
+        current.ShouldDelete = false
+        if inner != nil && current.Error == nil {
+            current.Error = inner
+        }
+    }
+    return current
 }
 
 func main() {
@@ -152,11 +153,11 @@ func main() {
 	sqsClient := sqs.NewFromConfig(cfg)
 
 	// Optionally install a custom policy that forces retry for handler errors.
-	var opts []sqsrouter.RouterOption
-	if os.Getenv("E2E_POLICY_FORCE_RETRY_ON_HANDLER_ERR") == "1" {
-		// Custom policy: turn any handler error into a retry (ShouldDelete=false)
-		opts = append(opts, sqsrouter.WithPolicy(forceRetryOnHandlerErr{}))
-	}
+    var opts []sqsrouter.RouterOption
+    if os.Getenv("E2E_POLICY_FORCE_RETRY_ON_HANDLER_ERR") == "1" {
+        // Custom policy: turn any handler error into a retry (ShouldDelete=false)
+        opts = append(opts, sqsrouter.WithPolicy(forceRetryOnHandlerErr{}))
+    }
 
 	router, err := sqsrouter.NewRouter(sqsrouter.EnvelopeSchema, opts...)
 	if err != nil {
@@ -179,8 +180,8 @@ func main() {
 
 	router.Register(MsgTypeE2ETest, MsgVersion1_0, E2ETestHandler)
 
-	consumer := sqsrouter.NewConsumer(sqsClient, queueURL, router)
-	consumer.Start(appCtx)
+    c := consumer.NewConsumer(sqsClient, queueURL, router)
+    c.Start(appCtx)
 
 	log.Println("Application has shut down.")
 }
